remove version range support from `remove_parameter_value_pairs()`
